+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## Version 1.3.1 - 2023-07-22
+
+### Fixed
+
+- Fixed a bug where parameter groups were omitted when using `TransformerBasedClassification`'s layer-specific fine-tuning functionality. ([#36](https://github.com/webis-de/small-text/issues/36), [#38](https://github.com/webis-de/small-text/pull/38))
+- Fixed a bug where class weighting resulted in `nan` values. ([#39](https://github.com/webis-de/small-text/issues/39))
+
+### Contributors
+
+[@JP-SystemsX](https://github.com/JP-SystemsX)
+
+---
+
 ## Version 1.3.0 - 2023-02-21
 
 ### Added
@@ -21,6 +34,8 @@
 ### Contributors
 
 [@rmitsch](https://github.com/rmitsch)
+
+---
 
 ## Version 1.2.0 - 2023-02-04